requirement.txt
# Python 版本3.12

# LangChain 主包
langchain>=0.3.0

# 拆分出来的官方集成
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0

# 向量数据库
chromadb>=0.5.0

# 配置文件 / YAML
PyYAML>=6.0.0